Remove simple and unused unicode conversion functions (v2)
The functions to convert to/from Unicode are too simple to be usable for any
application: they only work for ASCII and ISO-8859-1. This is usually better
done by the application with functions offered by Glib, Qt, iconv or similar.
The example applications are not converted to alternatives, instead they get a
private copy of those functions.