Make the IrDA device enumeration accessable to the application
Currently, the IrDA transport can only use the first device. It already uses a
device search to determine possible candidates. These changes make these search
results available to the application by the same functions that are used for
USB interface search. To make that possible, quite some parts of the irobex
transport are reworked and the transport connect function is now a wrapper
around the public interface. This also adds support for IAS queries on Windows.