Add read callback for custom transport
Currently, custom transport need their own buffer of the size of the largest
obex packet to do the custom data feed. This includes one data copy that can
be avoided by defining a read function like for all other transports. The old
method can be used at the same time.
Application must be recompiled! But we did not release, yet, and the next
release will have a different SOVERSION, so no problem.